Add action support to the Stream Deck plugin
- A new button type is now registered with the Stream Deck plugin for arbitrary Actions.
- Minor refactoring of a few things to keep things looking clean and pretty.

The code contained in the JS files is becoming more and more broken as time goes on. I plan to rewrite all of it in TypeScript and to be much better, but this lets me iterate rapidly and test things locally, sooooo... we get to live with it for now. I will note that I am ashamed of this code.