Initial Web API support
This commit adds initial support for a web-based API via EmbedIO. This API is very experimental and will be subject to change, and may not even actually ever go anywhere depending on how testing goes.

For now, the old websocket API is still active and fully functional to allow for backwards compatibility while things are being tested.

- Add API routes for:
  - Actions (/action)
  - Classes (/classes)
  - Hotbars (/hotbar)
  - Icons (/icon)
- Add EmbedIO websocket support